Removed unnecessary enable_if.
The variadic ctors that forward arguments to the event_attributes
constructor were unnecessarily disabled for when the first argument was
an event_attributes to avoid a copy.
I don't know why I did this originally as it is not necessary.

When a concrete `event_attributes` object is passed, the overload that
expects an `event_attributes` will be unambiguously selected.