Add overload for mark() to support directly passing attributes
Also change example in comments for mark() from lock/unlock of a mutex to discovering a problem.  NVTX has lock/unlock mutex labeling in the C API and we will add C++ wrappers for that soon, so this change prevents any future confusion from saying mark() is good for labeling locks.