Use constexpr in C++11 where possible
Changed a few places in nvtx3.hpp were using the NVTX3_CONSTEXPR_IF_CPP14 macro where a normal "constexpr" would work.  The macro is only needed in constructors with code in the body.  A constructor with a member initializer list and then empty body can be constexpr in C++11.  This improves codegen in GCC 4.x, VS2015, or using any compiler in -std:c++11 mode.