Update CMake files for C/C++ API
Add support for IMPORTED targets and version checks when using more
than one instance of the NVTX library with CMake.

Move most content from CMakeLists.txt into nvtxImportedTargets.cmake.
By default, nvtxImportedTargets.cmake defines the targets IMPORTED.

Use "include(nvtxImportedTargets.cmake)" in a library's CMake code
to define the NVTX targets locally-scoped to the project, ensuring
the correct NVTX version is used.

Use "add_subdirectory" on the NVTX c directory to use CMakeLists.txt,
which includes nvtxImportedTargets.cmake, but changes the targets to
not be IMPORTED.  This beneficial when adding NVTX to a project-wide
imports directory and defining global targets for all projects to use.
In this usage, if a newer NVTX version is added after an older one,
first-one-wins, so a warning will be printed when running CMake that
some library is using an older-than-expected NVTX version.

Tested on Linux and Windows with a range of compilers and a variety
of project configurations using multiple imported copies of NVTX.