Add macro wrappers for maybe_unused and nodiscard
Some compilers issues spurious warnings about unused variables for
instances of scoped_range and unique_range. Use [[maybe_unused]] when available
to silence these warnings.

Likewise, added `[[nodiscard]]` when available to `start_range()` to issue a warning
if a user accidently forgets to store the result of `start_range()`. It is not possible to correctly
use `start_range()` without using the result to call `end_range()`.