NVTX C++ - avoid using SFINAE in GCC 8.x < 8.4, due to known compiler bug
NCCL and QNX SDP 7.1 both use GCC 8.3, and we found that GCC 8.3 had a
bug that caused SFINAE techniques to trigger compile errors in code that
worked fine in all other GCC versions.  The bug was fixed in GCC 8.4.
See: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=87748

MSVC versions before 1914 (2017 15.7) had a similar problem.  NVTX C++
provides the macro NVTX3_USE_CHECKED_OVERLOADS_FOR_GET to choose whether
or not to use SFINAE in order to have a workaround for these cases.
That macro can be overridden by users, but should have good defaults to
avoid SFINAE on compilers that are known to not support it correctly.
Previously, only MSVC versions before 1914 defaulted this macro to off.
This change adds GCC 8 versions before 8.4 to the condition to default
the macro to off.

Tested manually with GCC 8.3 and 8.4, in cases expected to pass and fail.
Confirmed errors occur only in cases expected to fail, e.g. a domain
struct with no 'name' member, and error messages are the SFINAE-based
static_assert in 8.4 and a standard compile error in 8.3.