nvtx3.hpp: Add support for extended payload data
This adds a payload_data class and respective
handlers for event_attributes. The schema class
and some helper macros take care of wrapping the
schema id management.

For testing of the deleted rvalue construction, a
"ShouldFail" test facility is instroduced. These
tests are based on trying to compile code snippets
and expecting failure during compilation.

Regular tests are also added.