nvtx3.hpp: Add hidden macro to enable rvalue construction
If NVTX3_ALLOW_RVALUE_CONSTRUCTORS is defined,
rvalue constructors of non-owning classes
are no longer deleted as a safety measure.

This could be dangerous, but it allows concise
chains of constructing and consuming attributes
right away, which is also safe.

This is an undocumented experimental feature with
no guarantee for future compatibility.