Python Tests: Injection module to be used by Python tests
The implementation does the minimal required by the tests.
I.e., it records every NVTX events that are used by the tests.

Since currently there are no multi-threading test cases, the library is thread unsafe.

Additionally, domain and strings handles are not saved per domain,
as it keeps the implementation simpler, and the tests validates the attributes of the registration events anyway.

Currently, only `NVTX_PAYLOAD_TYPE_INT64` and `NVTX_PAYLOAD_TYPE_DOUBLE`
payload types are supported by the injection library.
Support for extended payloads and test cases for numpy payloads
will be added in a different task.