Python: massive tests refactor
This change significantly increases the code coverage of the Python unit tests, by using a dedicated injection tool.

With this change, `pytest` command runs the tests with the injection tool if the `--enable-injection` flag was provided.

The refactoring also includes:

- Use the same parameterized arguments (domain, message, etc...)  in all test cases.
  So, `test_categories_basic` is removed since it is covered by other tests.

- Refactor `test_profiler` to be pure Python, rather than Cython test.
  So, `python/tests/cython` directory is removed.

- Removed `test_nvtx_annotate_pushes_and_pops.py` since mocks are not needed anymore.
  The test cases for "annotate with exception" were moved to test_basic.py

- Refactor `test_pickle_annotate` to validate pickling annotated function
rather than context manager, as it is a more realistic use-case.